Create Kind::Checker::Factory and Kind::Checker::Protocol
